feat: Migrate contact form request to tRPC
